requirements.txt

fastapi
uvicorn

mistralai
google-generativeai
transformers
requests